switch to float32

+++ bezier.go
@@ -2,7 +2,7 @@ package gg
 
 import "math"
 
-func quadratic(x0, y0, x1, y1, x2, y2, t float64) (x, y float64) {
+func quadratic(x0, y0, x1, y1, x2, y2, t float32) (x, y float32) {
 	u := 1 - t
 	a := u * u
 	b := 2 * u * t
@@ -12,24 +12,24 @@ func quadratic(x0, y0, x1, y1, x2, y2, t float64) (x, y float64) {
 	return
 }
 
-func QuadraticBezier(x0, y0, x1, y1, x2, y2 float64) []Point {
-	l := (math.Hypot(x1-x0, y1-y0) +
-		math.Hypot(x2-x1, y2-y1))
+func QuadraticBezier(x0, y0, x1, y1, x2, y2 float32) []Point {
+	l := (math.Hypot(float64(x1-x0), float64(y1-y0)) +
+		math.Hypot(float64(x2-x1), float64(y2-y1)))
 	n := int(l + 0.5)
 	if n < 4 {
 		n = 4
 	}
-	d := float64(n) - 1
+	d := float32(n) - 1
 	result := make([]Point, n)
 	for i := 0; i < n; i++ {
-		t := float64(i) / d
+		t := float32(i) / d
 		x, y := quadratic(x0, y0, x1, y1, x2, y2, t)
 		result[i] = Point{x, y}
 	}
 	return result
 }
 
-func cubic(x0, y0, x1, y1, x2, y2, x3, y3, t float64) (x, y float64) {
+func cubic(x0, y0, x1, y1, x2, y2, x3, y3, t float32) (x, y float32) {
 	u := 1 - t
 	a := u * u * u
 	b := 3 * u * u * t
@@ -40,18 +40,18 @@ func cubic(x0, y0, x1, y1, x2, y2, x3, y3, t float64) (x, y float64) {
 	return
 }
 
-func CubicBezier(x0, y0, x1, y1, x2, y2, x3, y3 float64) []Point {
-	l := (math.Hypot(x1-x0, y1-y0) +
-		math.Hypot(x2-x1, y2-y1) +
-		math.Hypot(x3-x2, y3-y2))
+func CubicBezier(x0, y0, x1, y1, x2, y2, x3, y3 float32) []Point {
+	l := (math.Hypot(float64(x1-x0), float64(y1-y0)) +
+		math.Hypot(float64(x2-x1), float64(y2-y1)) +
+		math.Hypot(float64(x3-x2), float64(y3-y2)))
 	n := int(l + 0.5)
 	if n < 4 {
 		n = 4
 	}
-	d := float64(n) - 1
+	d := float32(n) - 1
 	result := make([]Point, n)
 	for i := 0; i < n; i++ {
-		t := float64(i) / d
+		t := float32(i) / d
 		x, y := cubic(x0, y0, x1, y1, x2, y2, x3, y3, t)
 		result[i] = Point{x, y}
 	}
