+++ examples/ellipse.go
@@ -8,7 +8,7 @@ func main() {
 	dc.SetRGBA(0, 0, 0, 0.1)
 	for i := 0; i < 360; i += 15 {
 		dc.Push()
-		dc.RotateAbout(gg.Radians(float64(i)), S/2, S/2)
+		dc.RotateAbout(gg.Radians(float32(i)), S/2, S/2)
 		dc.DrawEllipse(S/2, S/2, S*7/16, S/8)
 		dc.Fill()
 		dc.Pop()
